# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles:
    post:
      operationId: upsertProfile
      tags: [profiles]
      security:
        - verificationToken: []
      summary: Create or update a user profile
      description: >
        Creates or updates the caller's profile. Caller identity is derived
        server-side from the verification token issued by POST /auth/verify-otp.
        Pass the token as "Authorization: Bearer <token>".
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProfileRequest"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResult"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "401":
          description: Missing or invalid verification token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
  /profiles/location:
    post:
      operationId: updateLocation
      tags: [profiles]
      security:
        - verificationToken: []
      summary: Update the caller's current location
      description: >
        Stores the caller's latest latitude/longitude. Caller identity is derived
        server-side from the verification token. The profile must already exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationRequest"
      responses:
        "200":
          description: Location updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationResult"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "401":
          description: Missing or invalid verification token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
  /profiles/nearby:
    get:
      operationId: getNearbyProfiles
      tags: [profiles]
      security:
        - verificationToken: []
      summary: Get profiles within ~30 metres of the caller
      description: >
        Returns profiles whose last known location is within 30 metres of the
        caller's stored location. The caller is excluded from results.
        Each profile card includes a short AI-generated conversation-starter summary.
        Caller identity is derived server-side from the verification token.
      responses:
        "200":
          description: List of nearby profiles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NearbyProfilesResult"
        "400":
          description: Caller has no stored location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "401":
          description: Missing or invalid verification token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
        "404":
          description: Caller profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"
components:
  securitySchemes:
    verificationToken:
      type: http
      scheme: bearer
      description: >
        Short-lived opaque token issued by POST /auth/verify-otp.
        Valid for 30 minutes. Pass as "Authorization: Bearer <token>".
  schemas:
...
    ErrorResult:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UpsertProfileRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          description: Display name
        about:
          type: string
          description: Short bio / about text
        photo:
          type: string
          description: Profile photo as a data URL
    ProfileResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    UpdateLocationRequest:
      type: object
      required: [latitude, longitude]
      properties:
        latitude:
          type: number
          description: WGS-84 latitude in decimal degrees
        longitude:
          type: number
          description: WGS-84 longitude in decimal degrees
    LocationResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
    NearbyProfileCard:
      type: object
      required: [name, photo, distanceMeters, conversationStarter]
      properties:
        name:
          type: string
        photo:
          type: string
          description: Profile photo data URL
        distanceMeters:
          type: number
          description: Great-circle distance from the requester in metres
        conversationStarter:
          type: string
          description: Short AI-generated conversation-starter line based on the person's bio
    NearbyProfilesResult:
      type: object
      required: [profiles]
      properties:
        profiles:
          type: array
          items:
            $ref: "#/components/schemas/NearbyProfileCard"