# lib/api-zod/src/generated/types/nearbyProfileCard.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NearbyProfileCard {
  name: string;
  /** Profile photo data URL */
  photo: string;
  /** Great-circle distance from the requester in metres */
  distanceMeters: number;
  /** Short AI-generated conversation-starter line based on the person's bio */
  conversationStarter: string;
}